[#60] Handle exceptions in Chain operator
Previously, if the function passed to the `Chain` operator threw an exception, it would crash the `FlowEngine`.

This commit wraps the invocation of the chaining function in a `try...catch` block, ensuring that any thrown exception is caught and returned as a `Failure` outcome.

A new unit test has been added to verify this behavior.