[#21] Implement initial AST rewriting for WithRetry
This commit contains the first pass at implementing the AST rewriting
approach for the WithRetry operator.

It introduces the IBehaviourStrategy interface and a RetryStrategy
that handles the CreateNode case. The FlowExtensions have been updated
to use this new pattern.

NOTE: This implementation is incomplete and contains known flaws that
will be addressed in subsequent commits.